Update Neo4j example:  - move the example from examples/integrations/neo4j/neo4j-se to examples/integrations/neo4j  - renamed package from io.helidon.examples.integrations.neo4j.se to io.helidon.examples.integrations.neo4j  - update README  - fix metrics
Updated module-info and poms to remove CDI dependencies.

Adjust copyright years.
Update README to state JDK20 instead of JDK21.
