+++ etc/copyright-exclude.txt
@@ -47,10 +47,10 @@ node_modules
 archetype-resources/pom.xml
 # excluded as it contains both Oracle and original copyright notice
 src/main/java/org/jboss/weld/bean/proxy/ProxyFactory.java
-src/main/java/io/helidon/examples/integrations/neo4j/se/domain/Actor.java
-src/main/java/io/helidon/examples/integrations/neo4j/se/domain/Movie.java
-src/main/java/io/helidon/examples/integrations/neo4j/se/domain/Person.java
-src/main/java/io/helidon/examples/integrations/neo4j/se/domain/MovieRepository.java
+src/main/java/io/helidon/examples/integrations/neo4j/domain/Actor.java
+src/main/java/io/helidon/examples/integrations/neo4j/domain/Movie.java
+src/main/java/io/helidon/examples/integrations/neo4j/domain/Person.java
+src/main/java/io/helidon/examples/integrations/neo4j/domain/MovieRepository.java
 etc/THIRD_PARTY_LICENSES.xml
 etc/HELIDON_THIRD_PARTY_LICENSES.xml
 etc/dependency-check-suppression.xml
