--- examples/integrations/neo4j/neo4j-se/.dockerignore
@@ -1,1 +1,0 @@
-target/*
