+++ examples/integrations/neo4j/src/main/java/io/helidon/examples/integrations/neo4j/domain/Movie.java
@@ -1,0 +1,102 @@
+/*
+ * Copyright (c) 2002-2020 "Neo4j,"
+ * Neo4j Sweden AB [http://neo4j.com]
+ * This file is part of Neo4j.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.helidon.examples.integrations.neo4j.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/*
+ * Helidon changes are under the copyright of:
+ *
+ * Copyright (c) 2021, 2023 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+/**
+ * The Movie class.
+ *
+ * @author Michael Simons
+ */
+public class Movie {
+
+    private final String title;
+
+    private final String description;
+
+    private List<Actor> actors = new ArrayList<>();
+
+    private List<Person> directors = new ArrayList<>();
+
+    private Integer released;
+
+    /**
+     * Constructor for Movie.
+     *
+     * @param title
+     * @param description
+     */
+    public Movie(String title, String description) {
+        this.title = title;
+        this.description = description;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public List<Actor> getActors() {
+        return actors;
+    }
+
+    public void setActors(List<Actor> actors) {
+        this.actors = actors;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public List<Person> getDirectors() {
+        return directors;
+    }
+
+    public void setDirectorss(List<Person> directors) {
+        this.directors = directors;
+    }
+
+    public Integer getReleased() {
+        return released;
+    }
+
+    public void setReleased(Integer released) {
+        this.released = released;
+    }
+}
