+++ integrations/neo4j/health/src/main/java/module-info.java
@@ -29,10 +29,10 @@ import io.helidon.common.features.api.Preview;
 )
 module io.helidon.integrations.neo4j.health {
     requires static io.helidon.common.features.api;
+    requires io.helidon.health;
 
     requires org.neo4j.driver;
 
-    requires io.helidon.health;
 
     exports io.helidon.integrations.neo4j.health;
