+++ integrations/neo4j/metrics/pom.xml
@@ -52,12 +52,6 @@
             <artifactId>helidon-common-features-api</artifactId>
             <optional>true</optional>
         </dependency>
-        <dependency>
-            <!-- only needed for compilation, not required in runtime -->
-            <groupId>jakarta.enterprise</groupId>
-            <artifactId>jakarta.enterprise.cdi-api</artifactId>
-            <optional>true</optional>
-        </dependency>
     </dependencies>
 
     <build>
