+++ integrations/neo4j/neo4j/pom.xml
@@ -49,12 +49,6 @@
             <optional>true</optional>
         </dependency>
         <dependency>
-            <!-- only needed for compilation, not required in runtime -->
-            <groupId>jakarta.enterprise</groupId>
-            <artifactId>jakarta.enterprise.cdi-api</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
             <groupId>io.helidon.config</groupId>
             <artifactId>helidon-config-mp</artifactId>
             <optional>true</optional>
